CBD-4350: Revamp implementation of "unarchive" directive
This change canonicalizes the previous convention that "cbdep install
<package> <version>" will put the package in a directory named exactly
"<install_dir>/<package>-<version>", which we call the "target
directory".

It also enhances the "unarchive" directive to add three new
sub-directives:

 - toplevel_dir <dir>: specifies the single subdirectory inside the
   downloaded archive which should be considered the payload. This
   directory will be renamed to the target directory. If not specified,
   cbdep assumes there's a payload directory with the exact right name
   <package>-<version> already. Replaces the old "rename_dir" directive.
 - no_toplevel_dir: specifies that the archive has no top-level
   directory, and that the target directory should be created and the
   entire contents of the archive should be placed there. Replaces the
   old "add_dir" directive.
 - target_dir <dir>: specifies an alternate naming convention for the
   target directory. Should not be used, but unfortunately there's now a
   long history of Golang being put into a target directory named
   "go<version>" without the hyphen. That is the only package which
   should use this directive, and at least it's now explicit in the
   config file.

By making the target directory naming explicit and clarifying that the
various other directives apply exclusively to unarchiving, this change
allowed me to implement safer unarchiving. Now the downloaded URL is
always unarchived into a temp directory, then quickly renamed to the
target directory. If the target directory already exists, it is renamed
away first (after the unarchive is done), and then deleted afterwards.
This makes the actual populating of the target directory very close to
atomic, meaning virtually no chance of partially-extracted archives.

This change does also make explicit that "cbdep install" will create a
new directory, at least when "unarchive" is used. Previously some
packages would merge into an existing target directory, or even throw
errors if the target existed. Because this is a potentially noticeable
change in behaviour, I'm bumping cbdep to 1.1.0.

A further nice consequence is that we could remove some cleanup steps
from cbdep.config, in particular for the various JDK packages on Mac,
which previously had to spurious directories that were also at the top
level of the downloaded archives.

Note: the only old package that didn't easily fall into the above rules
was "bat_to_exe_converter". I've discovered that that package doesn't
even seem to exist anymore, so clearly it's been broken for some time
anyway. We haven't used it for anything in a long time, if we ever did.
I just removed it from cbdep.config.

Change-Id: I101541012aa9d4ba79e7cfc98da339afbdb2d58e
Reviewed-on: http://review.couchbase.org/c/cbdep/+/161574
Reviewed-by: Blair Watt <blair.watt@couchbase.com>
Tested-by: Chris Hillery <ceej@couchbase.com>