CBD-4352: Force Python's UTF-8 mode; Improved logging
After much investigation, the cause of the "ascii" decoding error on
recent golang tarballs when LANG is unset turned out to be a bug in
PyInstaller itself, which was incorrectly disabling Python's "utf8"
mode. There is a very recent fix for this on PyInstaller's develop
branch, so I have forked that, added it to the manifest, and now we
build a fresh PyInstaller each time rather than depending on the
packages from pip. This also required updating to Python 3.7+ on the
Centos6 build agent.

We still use PyInstaller 4.2 on MacOS x86_64, because newer versions
evidently attempt to auto-notarize binaries, and that isn't working. To
be solved another time.

Also fixed up cbdep's logging, and set it to INFO by default so that the
--debug option actually has meaning.

Change-Id: Iaeef6e03b2177a3f7ee7ec8bd3eefc7cd394e3e0
Reviewed-on: http://review.couchbase.org/c/cbdep/+/161661
Reviewed-by: Ming Ho <ming.ho@couchbase.com>
Tested-by: Chris Hillery <ceej@couchbase.com>