[wrappers] Use uv tool --python-preference=only-managed
`uv tool install` has had a series of bugs/unexpected behaviours when
choosing which python version to use, in particular when run in an
environment with an older Python on the PATH. Use
--python-preference=only-managed to tell uv to always download its own
python, which appears to avoid the issues.

Also add a Cloudfront invalidation step to the release-wrappers.sh
script, so that newly-released wrappers are immediately available to
users.

Change-Id: I4a73e3169769727220f2305f9697d59ae9a7e47b
Reviewed-on: https://review.couchbase.org/c/cbdep/+/223710
Tested-by: Chris Hillery <ceej@couchbase.com>
Reviewed-by: Ming Ho <ming.ho@couchbase.com>