CBD-6314: [1.3.4] Fix nodejs downloads on Alpine
Config blocks are fully checked in the order they're defined in
`cbdep.config`. For `nodejs`, the general `linux` block was listed first
before the `alpine` block. Because the list of architectures cbdep uses
includes `x64` on Alpine, this general `linux` block was matching on
Alpine, causing the non-MUSL variant of nodejs to be downloaded.

While fixing this, I reworked the algorithm used by `if_platform` and
`if_arch` to ensure that values in those directives are also checked in
the order specified. Previously, the algorithm would check each
introspected system value in turn against the list of values in the
directive. This meant that the introspected order dominated. With this
change, `cbdep.config` authors have more flexibility to specify the
platform and arch values from "most to least specific" without concern
for the implementation details of how platform introspection works.

I did also change `platform_introspection` to ensure that the `-musl`
and `-alpine` arch variants are all enumerated before any general arches
such as `x64`. This is no longer necessary due to the preceeding change
to `if_arch`, but it made sense to have it consistent for clarity.

Added a test case which initially failed for NodeJS Alpine downloads,
and now succeeds. Verified that no other test cases break with these
changes. Also modified a couple tests to emulate a change in
arch-introspection ordering, and verified that this doesn't break the
tests.

Change-Id: Ic3b76bde65595bf25413c1c68411a0fdd00b1c9f
Reviewed-on: https://review.couchbase.org/c/cbdep/+/227313
Reviewed-by: Ming Ho <ming.ho@couchbase.com>
Tested-by: Chris Hillery <ceej@couchbase.com>